pyproject.toml
[tool.poetry]
name = "martini-bot"
version = "0.1.0"
description = ""
authors = ["Remco Kersten"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
langchain = "^0.2.6"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.6"
streamlit = "^1.36.0"
openai = "^1.35.7"
tiktoken = "^0.7.0"
psutil = "^6.0.0"
langchain-chroma = "^0.1.2"
langchain-openai = "^0.1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
